PlayerRecordingLoader: Split methods into classes
