+++ Editor/PlayerRecordingLoaderEditor.cs
@@ -9,30 +9,32 @@ namespace HUMR
 {
     // Editor components prevent world build 
 #if UNITY_EDITOR
-    [CustomEditor(typeof(RecordLogLoader))]
-    public class RecordLogLoaderEditor : Editor
+    [CustomEditor(typeof(PlayerRecordingLoader))]
+    public class PlayerRecordingLoaderEditor : Editor
     {
         private bool _showAdvanced;
         
         public override void OnInspectorGUI()
         {
-            var recordLoader = (RecordLogLoader)target;
+            var recordLoader = (PlayerRecordingLoader)target;
             if (recordLoader == null) return;
-
+        
             UpdateLogDirectory(recordLoader);
             DrawAdvancedPathSection(recordLoader);
             
             InitializeLogs(recordLoader);
-
+        
             DrawLogFileDropdown(recordLoader);
             if (!DrawRecordingTargetDropdown(recordLoader)) return;
-
+        
+            GUILayout.Space(EditorGUIUtility.singleLineHeight);
+            
             recordLoader.exportGenericAnimation = GUILayout.Toggle(recordLoader.exportGenericAnimation, "Export Generic Animation");
             
             DrawExportButton(recordLoader);
         }
-
-        private void UpdateLogDirectory(RecordLogLoader recordLoader)
+        
+        private void UpdateLogDirectory(PlayerRecordingLoader recordLoader)
         {
             if (_showAdvanced) return;
             
@@ -42,16 +44,16 @@ namespace HUMR
                 recordLoader.logFileDirectory = defaultPath;
             }
         }
-
-        private void DrawAdvancedPathSection(RecordLogLoader recordLoader)
+        
+        private void DrawAdvancedPathSection(PlayerRecordingLoader recordLoader)
         {
             _showAdvanced = EditorGUILayout.Foldout(_showAdvanced, "Advanced: Custom Log Path");
             if (!_showAdvanced) return;
-
+        
             EditorGUI.indentLevel++;
             EditorGUILayout.BeginHorizontal();
             
-            recordLoader.logFileDirectory = EditorGUILayout.TextField("OutputLogPath", recordLoader.logFileDirectory);
+            recordLoader.logFileDirectory = EditorGUILayout.TextField("Output Log Path (resets when closed)", recordLoader.logFileDirectory);
             
             if (GUILayout.Button("Explore", GUILayout.Width(100)))
             {
@@ -61,7 +63,7 @@ namespace HUMR
             EditorGUILayout.EndHorizontal();
             EditorGUI.indentLevel--;
         }
-
+        
         private static void OpenLogFolder(string path)
         {
             if (Directory.Exists(path))
@@ -78,32 +80,32 @@ namespace HUMR
                 UnityEngine.Debug.LogError($"Log path does not exist: {path}");
             }
         }
-
-        private static void InitializeLogs(RecordLogLoader recordLoader)
+        
+        private static void InitializeLogs(PlayerRecordingLoader recordLoader)
         {
             if (recordLoader.recordFileNames == null)
             {
                 recordLoader.CollectLogFiles();
                 recordLoader.CollectRecordings();
             }
-
+        
             if (recordLoader.recordings == null)
             {
                 recordLoader.CollectRecordings();
             }
         }
-
-        private static void DrawLogFileDropdown(RecordLogLoader recordLoader)
+        
+        private static void DrawLogFileDropdown(PlayerRecordingLoader recordLoader)
         {
             var controlRect = EditorGUILayout.GetControlRect(true, EditorGUIUtility.singleLineHeight);
-            var popupRect = EditorGUI.PrefixLabel(controlRect, new GUIContent("Record Log File"));
+            var popupRect = EditorGUI.PrefixLabel(controlRect, new GUIContent("Recording Log File"));
             
             if (IsContextClick(popupRect))
             {
                 recordLoader.CollectLogFiles();
                 recordLoader.CollectRecordings();
             }
-
+        
             if (recordLoader.recordFileNames != null && recordLoader.recordFileNames.Length > 0)
             {
                 EditorGUI.BeginChangeCheck();
@@ -124,8 +126,8 @@ namespace HUMR
                 EditorGUI.Popup(popupRect, 0, new string[] { noRecordsMessage });
             }
         }
-
-        private static bool DrawRecordingTargetDropdown(RecordLogLoader recordLoader)
+        
+        private static bool DrawRecordingTargetDropdown(PlayerRecordingLoader recordLoader)
         {
             var controlRect = EditorGUILayout.GetControlRect(true, EditorGUIUtility.singleLineHeight);
             var popupRect = EditorGUI.PrefixLabel(controlRect, new GUIContent("Recording Target"));
@@ -134,33 +136,33 @@ namespace HUMR
             {
                 recordLoader.CollectRecordings();
             }
-
+        
             if (recordLoader.recordings != null && recordLoader.recordings.Count > 0)
             {
                 var recordListStr = recordLoader.recordings
                     .Select(entry => $"{entry.type}: {entry.name}")
                     .ToArray();
-
+        
                 recordLoader.recordingIndex = Mathf.Clamp(recordLoader.recordingIndex, 0, recordListStr.Length - 1);
                 recordLoader.recordingIndex = EditorGUI.Popup(popupRect, recordLoader.recordingIndex, recordListStr);
                 return true;
             }
-
+        
             EditorGUI.Popup(popupRect, 0, new string[] { "Recording data is corrupted." });
             return false;
-
+        
         }
-
-        private static void DrawExportButton(RecordLogLoader recordLoader)
+        
+        private static void DrawExportButton(PlayerRecordingLoader recordLoader)
         {
-            if (!GUILayout.Button("LoadLogToExportAnim")) return;
+            if (!GUILayout.Button("Load recording and export .fbx")) return;
             
             if (recordLoader.TryGetComponent<RecordLogLoaderInterface>(out var receiver))
             {
                 receiver.LoadRecordingAndExportAnim();
             }
         }
-
+        
         private static bool IsContextClick(Rect rect)
         {
             return Event.current.type == EventType.MouseDown &&
