+++ README.md
@@ -4,14 +4,14 @@
 
 A complete rewrite of the codebase so new features can be implemented more easily.
 
-> [!WARNING]
-> Remove the old `HUMR OutputLogLoader` package and `Prefabs`, `ReadMe`, `Scenes` and `Scripts` in `Assets/HUMR` before importing.
-
 - Combine Recorder and OutputLogLoader packages
 - New OutputLog format: semicolons to separate different types
 - Remove duplicated code
 
 ## Installation
+
+> [!WARNING]
+> Remove `Packages/HUMR OutputLogLoader` and `Prefabs`, `ReadMe`, `Scenes` and `Scripts` in `Assets/HUMR` before importing.
 
 ### Requirements
