+++ Runtime/CSharp/AnimationAssetExporter.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 8340da9e81244bc38c4c9a0746e339a1
+timeCreated: 1782916982
