+++ Runtime/CSharp/AvatarPoseSnapshot.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d4b14af76c8a4bc6ac7fb3d5c424c3a7
+timeCreated: 1782916923
