+++ Runtime/CSharp/PathUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ffa3cecf1729e6d4681978195d0d6d3b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
