+++ Runtime/UdonSharp/BaseRecorder.cs
@@ -6,18 +6,20 @@ using VRC.SDKBase;
 
 namespace HUMR
 {
-    
     public class BaseRecorder : UdonSharpBehaviour
     {
-        [SerializeField, Tooltip("Start recording button, connect to StartRecording event.")]
+        [SerializeField] [Tooltip("Start recording button, connect to StartRecording event.")]
         private Button startRecordButton;
-        [SerializeField, Tooltip("Stop recording button, connect to StartRecording event.")]
+
+        [SerializeField] [Tooltip("Stop recording button, connect to StartRecording event.")]
         private Button stopRecordButton;
-        [SerializeField, Tooltip("Start recording immediately on scene load.")]
+
+        [SerializeField] [Tooltip("Start recording immediately on scene load.")]
         protected bool recordOnStart = true;
-        [SerializeField, Tooltip("Frames per second for recording.")]
+
+        [SerializeField] [Tooltip("Frames per second for recording.")]
         protected float recordFramerate = 30;
-        
+
         protected float RecordTime;
         protected RecordingType RecordType = RecordingType.Object;
         protected string TargetName = "Target";
@@ -25,7 +27,7 @@ namespace HUMR
         private bool _isRecording;
         private float _recordInterval;
         private float _nextRecordTime;
-        
+
         private const string VariableDelimiter = ";";
         private const string ComponentDelimiter = ",";
 
@@ -37,7 +39,7 @@ namespace HUMR
         private void Update()
         {
             if (!_isRecording) return;
-            
+
             RecordTime += Time.deltaTime;
             if (RecordTime < _nextRecordTime) return;
             _nextRecordTime = RecordTime + _recordInterval;
@@ -54,7 +56,7 @@ namespace HUMR
             _isRecording = true;
             UpdateUI();
         }
-        
+
         public virtual void StopRecording()
         {
             RecordStop(RecordType, TargetName);
@@ -68,7 +70,9 @@ namespace HUMR
             if (stopRecordButton != null) stopRecordButton.gameObject.SetActive(_isRecording);
         }
 
-        protected virtual void OnRecordTick() { }
+        protected virtual void OnRecordTick()
+        {
+        }
 
         private void OnDestroy()
         {
@@ -84,55 +88,57 @@ namespace HUMR
         protected static void RecordObjectTransform(Transform transform, string name, float time)
         {
             var timeStr = time.ToString(CultureInfo.InvariantCulture);
-    
+
             var positionStr = FormatVector3Components(transform.position);
             var rotationStr = FormatQuaternionComponents(transform.rotation);
             var scaleStr = FormatVector3Components(transform.localScale);
 
             var outputString = string.Join(VariableDelimiter, name, timeStr, positionStr, rotationStr, scaleStr);
-    
-            HumrUtils.HumrLog(outputString);
+
+            RecorderUtils.HumrLog(outputString);
         }
 
         protected static void RecordPlayerBones(VRCPlayerApi player, float time)
         {
             var timeStr = time.ToString(CultureInfo.InvariantCulture);
-            
+
             var hipsPosition = player.GetBonePosition(HumanBodyBones.Hips);
             var hipsPositionStr = FormatVector3Components(hipsPosition);
-            
+
             var outputString = string.Join(VariableDelimiter, player.displayName, timeStr, hipsPositionStr);
-            
+
             for (var i = 0; i < (int)HumanBodyBones.LastBone; i++)
             {
                 var rotation = player.GetBoneRotation((HumanBodyBones)i);
                 var rotationStr = FormatQuaternionComponents(rotation);
                 outputString = string.Join(VariableDelimiter, outputString, rotationStr);
             }
-            
-            HumrUtils.HumrLog(outputString);
+
+            RecorderUtils.HumrLog(outputString);
         }
 
         private static string FormatVector3Components(Vector3 vector3)
         {
-            var trimmedVector3 = vector3.ToString().Trim('(',')');
+            var trimmedVector3 = vector3.ToString().Trim('(', ')');
             return trimmedVector3.Replace(" ", "");
         }
 
         private static string FormatQuaternionComponents(Quaternion quaternion)
         {
-            var trimmedQuaternion = quaternion.ToString().Trim('(',')');
+            var trimmedQuaternion = quaternion.ToString().Trim('(', ')');
             return trimmedQuaternion.Replace(" ", "");
         }
 
         private static void RecordStart(RecordingType recType, string recName)
         {
-            HumrUtils.HumrLog(string.Join(VariableDelimiter, HumrUtils.RecordingStarted, HumrUtils.RecTypeToString(recType), recName));
+            RecorderUtils.HumrLog(string.Join(VariableDelimiter, RecorderUtils.RecordingStarted,
+                RecorderUtils.RecTypeToString(recType), recName));
         }
 
         private static void RecordStop(RecordingType recType, string recName)
         {
-            HumrUtils.HumrLog(string.Join(VariableDelimiter, HumrUtils.RecordingStopped, HumrUtils.RecTypeToString(recType), recName));
+            RecorderUtils.HumrLog(string.Join(VariableDelimiter, RecorderUtils.RecordingStopped,
+                RecorderUtils.RecTypeToString(recType), recName));
         }
     }
 }
